requirements.txt
# Consolidated Python dependencies for AI Trading System
# Source: backend/requirements.txt
# Generated on 2025-12-06

# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# AI APIs
anthropic==0.8.0

# Async & HTTP
aiohttp==3.9.1
httpx==0.25.2
asyncio==3.4.3

# Database & Caching
redis==5.0.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
# psycopg removed - using asyncpg only for PostgreSQL connections

# Vector Database (RAG)
pgvector==0.2.4
openai==1.6.1
tiktoken==0.5.2

# Data & Market Data
pandas==2.1.4
yfinance==0.2.33
APScheduler==3.10.4
tenacity==8.2.3


# Machine Learning (News Context Filter)
scikit-learn==1.3.2
numpy==1.26.4 
scipy==1.11.4

# PDF & Reporting
reportlab==4.0.7
pillow==10.1.0
matplotlib==3.8.2

# News & Web Scraping
newsapi-python==0.2.7
feedparser==6.0.10
beautifulsoup4==4.12.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging & Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
google-generativeai>=0.3.0
